requirements.txt
pip-autoremove>=0.9.0
flask>=0.11
flask-restful>=0.2.1
flask-sqlalchemy>=2.1
sqlalchemy-utils>=0.32.9
passlib>=1.6.5
python-jose>=0.5.2
pytest>=3.0.3
pyinstaller>=3.2
uwsgi>=2.0.14
requests>=2.11.1